Add game_state folder and Interface class
Refactored interface-related logic from main, including Screen and UI classes, to the game_state folder for better organization.

Note: Added empty "__init__.py" files in each folder (excluding the project root) to ensure Python can properly navigate the project structure.
